Change `ConditionallySelectable` supertrait
To resolve #94, removes the `Copy` supertrait bound on
`ConditionallySelectable`, replacing it with `Sized` instead.

It turns out the bound is only used in the default implementation of
`ConditionallySelectable::conditional_swap`, and is easy to replace by
slightly changing that default implementation.

Removing this supertrait bound is arguably a breaking change since it
means types which impl `ConditionallySelectable` can no longer be
assumed to be `Copy`, so also bumps the version to `3.0.0-pre`.